Move ChoiceGroup v8 changes to react
