--- apps/perf-test/src/scenarios/ChoiceGroupNext.tsx
@@ -1,13 +1,0 @@
-import * as React from 'react';
-import { ChoiceGroup, IChoiceGroupOption } from '@fluentui/react-next';
-
-const options: IChoiceGroupOption[] = [
-  { key: 'A', text: 'Option A' },
-  { key: 'B', text: 'Option B' },
-  { key: 'C', text: 'Option C', disabled: true },
-  { key: 'D', text: 'Option D' },
-];
-
-const Scenario = () => <ChoiceGroup defaultSelectedKey="B" options={options} label="Pick one" required={true} />;
-
-export default Scenario;
