--- packages/react-examples/src/react-next/ChoiceGroup/ChoiceGroup.Basic.Example.tsx
@@ -1,17 +1,0 @@
-import * as React from 'react';
-import { ChoiceGroup, IChoiceGroupOption } from '@fluentui/react-next/lib/ChoiceGroup';
-
-const options: IChoiceGroupOption[] = [
-  { key: 'A', text: 'Option A' },
-  { key: 'B', text: 'Option B' },
-  { key: 'C', text: 'Option C', disabled: true },
-  { key: 'D', text: 'Option D' },
-];
-
-export const ChoiceGroupBasicExample: React.FunctionComponent = () => {
-  return <ChoiceGroup defaultSelectedKey="B" options={options} onChange={_onChange} label="Pick one" required={true} />;
-};
-
-function _onChange(ev: React.FormEvent<HTMLInputElement>, option: IChoiceGroupOption): void {
-  console.dir(option);
-}
