--- packages/react-examples/src/react-next/ChoiceGroup/ChoiceGroup.Icon.Example.tsx
@@ -1,12 +1,0 @@
-import * as React from 'react';
-import { ChoiceGroup, IChoiceGroupOption } from '@fluentui/react-next/lib/ChoiceGroup';
-
-const options: IChoiceGroupOption[] = [
-  { key: 'day', text: 'Day', iconProps: { iconName: 'CalendarDay' } },
-  { key: 'week', text: 'Week', iconProps: { iconName: 'CalendarWeek' } },
-  { key: 'month', text: 'Month', iconProps: { iconName: 'Calendar' }, disabled: true },
-];
-
-export const ChoiceGroupIconExample: React.FunctionComponent = () => {
-  return <ChoiceGroup label="Pick one icon" defaultSelectedKey="day" options={options} />;
-};
