--- packages/react-examples/src/react-next/ChoiceGroup/docs/ChoiceGroupOverview.md
@@ -1,1 +1,0 @@
-Radio buttons (`ChoiceGroup`) let people select a single option from two or more choices.
