+++ packages/react-examples/src/react/ChoiceGroup/ChoiceGroup.doc.tsx
@@ -18,7 +18,7 @@ const ChoiceGroupIconExampleCode = require('!raw-loader!@fluentui/react-examples
 export const ChoiceGroupPageProps: IDocPageProps = {
   title: 'ChoiceGroup',
   componentName: 'ChoiceGroup',
-  componentUrl: 'https://github.com/microsoft/fluentui/tree/master/packages/react/src/components/ChoiceGroup',
+  componentUrl: 'https://github.com/microsoft/fluentui/tree/master/packages/react/src-internal/components/ChoiceGroup',
   examples: [
     {
       title: 'Basic ChoiceGroup',
