+++ packages/react-internal/src/components/ChoiceGroup/ChoiceGroup.base.tsx
@@ -1,247 +1,165 @@
 import * as React from 'react';
-
 import { Label } from '../../Label';
+import { classNamesFunction, find, getNativeProps, divProperties, IRefObject } from '../../Utilities';
 import {
-  initializeComponentRef,
-  warnDeprecations,
-  warnMutuallyExclusive,
-  classNamesFunction,
-  find,
-  getId,
-  isControlled,
-  getNativeProps,
-  divProperties,
-} from '../../Utilities';
-import {
-  IChoiceGroup,
   IChoiceGroupOption,
   IChoiceGroupProps,
   IChoiceGroupStyleProps,
   IChoiceGroupStyles,
+  IChoiceGroup,
 } from './ChoiceGroup.types';
-import { ChoiceGroupOption, IChoiceGroupOptionProps } from './ChoiceGroupOption/index';
+import { IChoiceGroupOptionProps } from './ChoiceGroupOption/ChoiceGroupOption.types';
+import { ChoiceGroupOption } from './ChoiceGroupOption/index';
+import { useId, useControllableValue, useWarnings } from '@uifabric/react-hooks';
 
 const getClassNames = classNamesFunction<IChoiceGroupStyleProps, IChoiceGroupStyles>();
 
-export interface IChoiceGroupState {
-  /**
-   * Current selected option, for **internal use only**.
-   * External users should access `IChoiceGroup.checkedOption` instead.
-   */
-  // TODO (Fabric 8?) - once we removed the checked property from individual options,
-  // we can probably store only the uncontrolled value in the state (right now it tracks
-  // the value regardless of controlled/uncontrolled--though if controlled, it only updates
-  // the value in state when the selectedKey prop updates)
-  keyChecked?: string | number;
+const getOptionId = (option: IChoiceGroupOption, id: string): string => {
+  return `${id}-${option.key}`;
+};
 
-  /** Is set when the control has focus. */
-  keyFocused?: string | number;
-}
+const getCheckedOption = (options: IChoiceGroupOption[], keyChecked: string | number) =>
+  find(options, (value: IChoiceGroupOption) => value.key === keyChecked);
 
-/**
- * {@docCategory ChoiceGroup}
- */
-export class ChoiceGroupBase extends React.Component<IChoiceGroupProps, IChoiceGroupState> implements IChoiceGroup {
-  private _id: string;
-  private _labelId: string;
-  private _focusCallbacks: { [key: string]: IChoiceGroupOptionProps['onFocus'] } = {};
-  private _changeCallbacks: { [key: string]: IChoiceGroupOptionProps['onBlur'] } = {};
+const useComponentRef = (
+  options: IChoiceGroupOption[],
+  keyChecked: string | number | undefined,
+  id: string,
+  componentRef?: IRefObject<IChoiceGroup>,
+) => {
+  React.useImperativeHandle(
+    componentRef,
+    () => ({
+      get checkedOption() {
+        return getCheckedOption(options, keyChecked!);
+      },
+      focus() {
+        const optionToFocus = getCheckedOption(options, keyChecked!) || options.filter(option => !option.disabled)[0];
+        const elementToFocus = optionToFocus && document.getElementById(getOptionId(optionToFocus, id));
 
-  constructor(props: IChoiceGroupProps) {
-    super(props);
+        if (elementToFocus) {
+          elementToFocus.focus();
+        }
+      },
+    }),
+    [options, keyChecked, id],
+  );
+};
 
-    initializeComponentRef(this);
+const COMPONENT_NAME = 'ChoiceGroup';
 
-    if (process.env.NODE_ENV !== 'production') {
-      warnDeprecations('ChoiceGroup', props, { onChanged: 'onChange' });
-      warnMutuallyExclusive('ChoiceGroup', props, {
-        selectedKey: 'defaultSelectedKey',
-      });
+export const ChoiceGroupBase: React.FunctionComponent<IChoiceGroupProps> = React.forwardRef<
+  HTMLDivElement,
+  IChoiceGroupProps
+>((props, forwardedRef) => {
+  const {
+    className,
+    theme,
+    styles,
+    options = [],
+    label,
+    required,
+    disabled,
+    name,
+    defaultSelectedKey,
+    componentRef,
+    onChange,
+  } = props;
+  const id = useId('ChoiceGroup');
+  const labelId = useId('ChoiceGroupLabel');
+
+  const divProps = getNativeProps<React.HTMLAttributes<HTMLDivElement>>(props, divProperties, [
+    'onChange',
+    'className',
+    'required',
+  ]);
+
+  const classNames = getClassNames(styles!, {
+    theme: theme!,
+    className,
+    optionsContainIconOrImage: options.some(option => !!(option.iconProps || option.imageSrc)),
+  });
+
+  const ariaLabelledBy = props.ariaLabelledBy || (label ? labelId : props['aria-labelledby']);
+
+  const [keyChecked, setKeyChecked] = useControllableValue(props.selectedKey, defaultSelectedKey);
+  const [keyFocused, setKeyFocused] = React.useState<string | number>();
+
+  useDebugWarnings(props);
+  useComponentRef(options, keyChecked, id, componentRef);
+
+  const onFocus = React.useCallback((ev?: React.FocusEvent<HTMLElement>, option?: IChoiceGroupOptionProps) => {
+    if (option) {
+      setKeyFocused(option.itemKey);
+      option.onFocus?.(ev);
     }
+  }, []);
 
-    const { defaultSelectedKey, options = [] } = props;
-    const validDefaultSelectedKey =
-      !_isControlled(props) &&
-      defaultSelectedKey !== undefined &&
-      options.some(option => option.key === defaultSelectedKey);
+  const onBlur = React.useCallback((ev: React.FocusEvent<HTMLElement>, option?: IChoiceGroupOptionProps) => {
+    setKeyFocused(undefined);
+    option?.onBlur?.(ev);
+  }, []);
 
-    this.state = {
-      keyChecked: validDefaultSelectedKey ? defaultSelectedKey : this._getKeyChecked(props),
-    };
-
-    this._id = getId('ChoiceGroup');
-    this._labelId = getId('ChoiceGroupLabel');
-  }
-
-  /**
-   * Gets the current checked option.
-   */
-  public get checkedOption(): IChoiceGroupOption | undefined {
-    const { options = [] } = this.props;
-    return find(options, (value: IChoiceGroupOption) => value.key === this.state.keyChecked);
-  }
-
-  public componentDidUpdate(prevProps: IChoiceGroupProps, prevState: IChoiceGroupState): void {
-    // Only update if a new props object has been passed in (don't care about state updates)
-    if (prevProps !== this.props) {
-      const newKeyChecked = this._getKeyChecked(this.props);
-      const oldKeyChecked = this._getKeyChecked(prevProps);
-
-      if (newKeyChecked !== oldKeyChecked) {
-        this.setState({
-          keyChecked: newKeyChecked,
-        });
+  const onOptionChange = React.useCallback(
+    (evt?: React.FormEvent<HTMLElement | HTMLInputElement>, option?: IChoiceGroupOptionProps) => {
+      if (!option) {
+        return;
       }
-    }
-  }
+      setKeyChecked(option.itemKey);
 
-  public render(): JSX.Element {
-    const { className, theme, styles, options = [], label, required, disabled, name } = this.props;
-    const { keyChecked, keyFocused } = this.state;
+      option.onChange?.(evt);
+      onChange?.(
+        evt,
+        find(options || [], (value: IChoiceGroupOption) => value.key === option.itemKey),
+      );
+    },
+    [onChange, options, setKeyChecked],
+  );
 
-    const divProps = getNativeProps<React.HTMLAttributes<HTMLDivElement>>(this.props, divProperties, [
-      'onChange',
-      'className',
-      'required',
-    ]);
-
-    const classNames = getClassNames(styles!, {
-      theme: theme!,
-      className,
-      optionsContainIconOrImage: options.some(option => !!(option.iconProps || option.imageSrc)),
-    });
-
-    const labelId = this._id + '-label';
-    const ariaLabelledBy = this.props.ariaLabelledBy || (label ? labelId : this.props['aria-labelledby']);
-
-    // TODO (Fabric 8?) - if possible, move `root` class to the actual root and eliminate
-    // `applicationRole` class (but the div structure will stay the same by necessity)
-    return (
-      // eslint-disable-next-line deprecation/deprecation
-      <div className={classNames.applicationRole} {...divProps}>
-        <div
-          className={classNames.root}
-          role="radiogroup"
-          {...(ariaLabelledBy && { 'aria-labelledby': ariaLabelledBy })}
-        >
-          {label && (
-            <Label className={classNames.label} required={required} id={labelId} disabled={disabled}>
-              {label}
-            </Label>
-          )}
-          <div className={classNames.flexContainer}>
-            {options.map((option: IChoiceGroupOption) => {
-              const innerOptionProps = {
-                ...option,
-                focused: option.key === keyFocused,
-                checked: option.key === keyChecked,
-                disabled: option.disabled || disabled,
-                id: this._getOptionId(option),
-                labelId: this._getOptionLabelId(option),
-                name: name || this._id,
-                required,
-              };
-
-              return (
-                <ChoiceGroupOption
-                  key={option.key}
-                  onBlur={this._onBlur}
-                  onFocus={this._onFocus(option.key)}
-                  onChange={this._onChange(option.key)}
-                  {...innerOptionProps}
-                />
-              );
-            })}
-          </div>
+  return (
+    <div className={classNames.root} {...divProps} ref={forwardedRef}>
+      <div role="radiogroup" {...(ariaLabelledBy && { 'aria-labelledby': ariaLabelledBy })}>
+        {label && (
+          <Label className={classNames.label} required={required} id={labelId} disabled={disabled}>
+            {label}
+          </Label>
+        )}
+        <div className={classNames.flexContainer}>
+          {options.map((option: IChoiceGroupOption) => {
+            return (
+              <ChoiceGroupOption
+                key={option.key}
+                itemKey={option.key}
+                {...option}
+                onBlur={onBlur}
+                onFocus={onFocus}
+                onChange={onOptionChange}
+                focused={option.key === keyFocused}
+                checked={option.key === keyChecked}
+                disabled={option.disabled || disabled}
+                id={getOptionId(option, id)}
+                labelId={option.labelId || `${labelId}-${option.key}`}
+                name={name || id}
+                required={required}
+              />
+            );
+          })}
         </div>
       </div>
-    );
-  }
+    </div>
+  );
+});
+ChoiceGroupBase.displayName = COMPONENT_NAME;
 
-  public focus() {
-    const { options = [] } = this.props;
-    const optionToFocus = this.checkedOption || options.filter(option => !option.disabled)[0];
-    const elementToFocus = optionToFocus && document.getElementById(this._getOptionId(optionToFocus));
-    if (elementToFocus) {
-      elementToFocus.focus();
-    }
-  }
-
-  private _onFocus(key: string) {
-    // This extra mess is necessary because React won't pass the `key` prop through to ChoiceGroupOption
-    if (!this._focusCallbacks[key]) {
-      this._focusCallbacks[key] = (
-        ev: React.FocusEvent<HTMLElement | HTMLInputElement>,
-        option: IChoiceGroupOption,
-      ) => {
-        this.setState({
-          keyFocused: key,
-        });
-      };
-    }
-    return this._focusCallbacks[key];
-  }
-
-  private _onBlur = (ev: React.FocusEvent<HTMLElement>, option: IChoiceGroupOption) => {
-    this.setState({
-      keyFocused: undefined,
+function useDebugWarnings(props: IChoiceGroupProps) {
+  if (process.env.NODE_ENV !== 'production') {
+    // eslint-disable-next-line react-hooks/rules-of-hooks -- build-time conditional
+    useWarnings({
+      name: COMPONENT_NAME,
+      props,
+      mutuallyExclusive: {
+        selectedKey: 'defaultSelectedKey',
+      },
     });
-  };
-
-  private _onChange(key: string) {
-    // This extra mess is necessary because React won't pass the `key` prop through to ChoiceGroupOption
-    if (!this._changeCallbacks[key]) {
-      this._changeCallbacks[key] = (
-        evt: React.FormEvent<HTMLElement | HTMLInputElement>,
-        option: IChoiceGroupOption,
-      ) => {
-        // eslint-disable-next-line deprecation/deprecation
-        const { onChanged, onChange } = this.props;
-
-        // Only manage state in uncontrolled scenarios.
-        if (!_isControlled(this.props)) {
-          this.setState({
-            keyChecked: key,
-          });
-        }
-
-        // Get the original option without the `key` prop removed
-        const originalOption = find(this.props.options || [], (value: IChoiceGroupOption) => value.key === key);
-
-        // TODO: onChanged deprecated, remove else if after 07/17/2017 when onChanged has been removed.
-        if (onChange) {
-          onChange(evt, originalOption);
-        } else if (onChanged) {
-          onChanged(originalOption!, evt);
-        }
-      };
-    }
-    return this._changeCallbacks[key];
   }
-
-  /**
-   * Returns `selectedKey` if provided, or the key of the first option with the `checked` prop set.
-   */
-  private _getKeyChecked(props: IChoiceGroupProps): string | number | undefined {
-    if (props.selectedKey !== undefined) {
-      return props.selectedKey;
-    }
-
-    const { options = [] } = props;
-    // eslint-disable-next-line deprecation/deprecation
-    const optionsChecked = options.filter((option: IChoiceGroupOption) => option.checked);
-    return optionsChecked[0] && optionsChecked[0].key;
-  }
-
-  private _getOptionId(option: IChoiceGroupOption): string {
-    return option.id || `${this._id}-${option.key}`;
-  }
-
-  private _getOptionLabelId(option: IChoiceGroupOption): string {
-    return option.labelId || `${this._labelId}-${option.key}`;
-  }
-}
-
-function _isControlled(props: IChoiceGroupProps): boolean {
-  return isControlled(props, 'selectedKey');
 }
