+++ packages/react-internal/src/components/ChoiceGroup/ChoiceGroupOption/ChoiceGroupOption.base.tsx
@@ -6,146 +6,136 @@ import {
   IChoiceGroupOptionStyleProps,
   IChoiceGroupOptionStyles,
 } from './ChoiceGroupOption.types';
-import { classNamesFunction, getNativeProps, inputProperties, css, initializeComponentRef } from '../../../Utilities';
-import { IProcessedStyleSet } from '../../../Styling';
-import { composeRenderFunction } from '@uifabric/utilities';
+import {
+  classNamesFunction,
+  getNativeProps,
+  inputProperties,
+  css,
+  composeRenderFunction,
+  getPropsWithDefaults,
+} from '../../../Utilities';
 
 const getClassNames = classNamesFunction<IChoiceGroupOptionStyleProps, IChoiceGroupOptionStyles>();
 
 const LARGE_IMAGE_SIZE = 71;
 
-/**
- * {@docCategory ChoiceGroup}
- */
-export class ChoiceGroupOptionBase extends React.Component<IChoiceGroupOptionProps, {}> {
-  public static defaultProps: Partial<IChoiceGroupOptionProps> = {
-    // This ensures default imageSize value doesn't mutate. Mutation can cause style re-calcuation.
-    imageSize: { width: 32, height: 32 },
-  };
+const DEFAULT_PROPS: Partial<IChoiceGroupOptionProps> = {
+  // This ensures default imageSize value doesn't mutate. Mutation can cause style re-calcuation.
+  imageSize: { width: 32, height: 32 },
+};
+type IChoiceGroupOptionPropsWithKey = IChoiceGroupOptionProps & { key: string };
 
-  private _classNames: IProcessedStyleSet<IChoiceGroupOptionStyles>;
+export const ChoiceGroupOptionBase: React.FunctionComponent<IChoiceGroupOptionProps> = propsWithoutDefaults => {
+  // Mix the `key` prop back in since custom render functions may be expecting it
+  const props = getPropsWithDefaults(
+    { ...DEFAULT_PROPS, key: propsWithoutDefaults.itemKey },
+    propsWithoutDefaults as IChoiceGroupOptionPropsWithKey,
+  );
 
-  constructor(props: IChoiceGroupOptionProps) {
-    super(props);
-    initializeComponentRef(this);
-  }
+  const {
+    ariaLabel,
+    focused,
+    required,
+    theme,
+    iconProps,
+    imageSrc,
+    imageSize,
+    disabled,
+    checked,
+    id,
+    styles,
+    name,
+    ...rest
+  } = props;
 
-  public render(): JSX.Element {
-    const {
-      ariaLabel,
-      focused,
-      required,
-      theme,
-      iconProps,
-      imageSrc,
-      imageSize,
-      disabled,
-      // eslint-disable-next-line deprecation/deprecation
-      checked,
-      id,
-      styles,
-      name,
-      onRenderField = this._onRenderField,
-      ...rest
-    } = this.props;
+  const classNames = getClassNames(styles!, {
+    theme: theme!,
+    hasIcon: !!iconProps,
+    hasImage: !!imageSrc,
+    checked,
+    disabled,
+    imageIsLarge: !!imageSrc && (imageSize!.width > LARGE_IMAGE_SIZE || imageSize!.height > LARGE_IMAGE_SIZE),
+    imageSize,
+    focused,
+  });
 
-    this._classNames = getClassNames(styles!, {
-      theme: theme!,
-      hasIcon: !!iconProps,
-      hasImage: !!imageSrc,
-      checked,
-      disabled,
-      imageIsLarge: !!imageSrc && (imageSize!.width > LARGE_IMAGE_SIZE || imageSize!.height > LARGE_IMAGE_SIZE),
-      imageSize,
-      focused,
-    });
+  const { className, ...nativeProps } = getNativeProps<{ className: string }>(rest, inputProperties);
 
-    const { className, ...nativeProps } = getNativeProps<{ className: string }>(rest, inputProperties);
-
-    return (
-      <div className={this._classNames.root}>
-        <div className={this._classNames.choiceFieldWrapper}>
-          <input
-            aria-label={ariaLabel}
-            id={id}
-            className={css(this._classNames.input, className)}
-            type="radio"
-            name={name}
-            disabled={disabled}
-            checked={checked}
-            required={required}
-            {...nativeProps}
-            onChange={this._onChange}
-            onFocus={this._onFocus}
-            onBlur={this._onBlur}
-          />
-          {onRenderField(this.props, this._onRenderField)}
-        </div>
-      </div>
-    );
-  }
-
-  private _onChange = (evt: React.FormEvent<HTMLInputElement>): void => {
-    const { onChange } = this.props;
-    if (onChange) {
-      onChange(evt, this.props);
-    }
-  };
-
-  private _onBlur = (evt: React.FocusEvent<HTMLElement>) => {
-    const { onBlur } = this.props;
-    if (onBlur) {
-      onBlur(evt, this.props);
-    }
-  };
-
-  private _onFocus = (evt: React.FocusEvent<HTMLElement>) => {
-    const { onFocus } = this.props;
-    if (onFocus) {
-      onFocus(evt, this.props);
-    }
-  };
-
-  private _onRenderField = (props: IChoiceGroupOptionProps): JSX.Element => {
-    const { id, imageSrc, imageAlt = '', selectedImageSrc, iconProps } = props;
-
-    const imageSize = props.imageSize ? props.imageSize : { width: 32, height: 32 };
-
-    const onRenderLabel = props.onRenderLabel
-      ? composeRenderFunction(props.onRenderLabel, this._onRenderLabel)
-      : this._onRenderLabel;
-
-    const label = onRenderLabel(props);
-
-    return (
-      <label htmlFor={id} className={this._classNames.field}>
-        {imageSrc && (
-          <div className={this._classNames.innerField}>
-            <div className={this._classNames.imageWrapper}>
-              <Image src={imageSrc} alt={imageAlt} width={imageSize.width} height={imageSize.height} />
-            </div>
-            <div className={this._classNames.selectedImageWrapper}>
-              <Image src={selectedImageSrc} alt={imageAlt} width={imageSize.width} height={imageSize.height} />
-            </div>
-          </div>
-        )}
-        {iconProps && (
-          <div className={this._classNames.innerField}>
-            <div className={this._classNames.iconWrapper}>
-              <Icon {...iconProps} />
-            </div>
-          </div>
-        )}
-        {imageSrc || iconProps ? <div className={this._classNames.labelWrapper}>{label}</div> : label}
-      </label>
-    );
-  };
-
-  private _onRenderLabel = (props: IChoiceGroupOptionProps): JSX.Element => {
+  const defaultOnRenderLabel = (): JSX.Element => {
     return (
       <span id={props.labelId} className="ms-ChoiceFieldLabel">
         {props.text}
       </span>
     );
   };
-}
+
+  const defaultOnRenderField = (): JSX.Element => {
+    const { imageAlt = '', selectedImageSrc } = props;
+
+    const onRenderLabel = props.onRenderLabel
+      ? composeRenderFunction(props.onRenderLabel, defaultOnRenderLabel)
+      : defaultOnRenderLabel;
+
+    const label = onRenderLabel(props);
+
+    return (
+      <label htmlFor={id} className={classNames.field}>
+        {imageSrc && (
+          <div className={classNames.innerField}>
+            <div className={classNames.imageWrapper}>
+              <Image src={imageSrc} alt={imageAlt} {...imageSize} />
+            </div>
+            <div className={classNames.selectedImageWrapper}>
+              <Image src={selectedImageSrc} alt={imageAlt} {...imageSize} />
+            </div>
+          </div>
+        )}
+        {iconProps && (
+          <div className={classNames.innerField}>
+            <div className={classNames.iconWrapper}>
+              <Icon {...iconProps} />
+            </div>
+          </div>
+        )}
+        {imageSrc || iconProps ? <div className={classNames.labelWrapper}>{label}</div> : label}
+      </label>
+    );
+  };
+
+  const { onRenderField = defaultOnRenderField } = props;
+
+  const onChange = (evt: React.FormEvent<HTMLInputElement>): void => {
+    props.onChange?.(evt, props);
+  };
+
+  const onBlur = (evt: React.FocusEvent<HTMLElement>) => {
+    props.onBlur?.(evt);
+  };
+
+  const onFocus = (evt: React.FocusEvent<HTMLElement>) => {
+    props.onFocus?.(evt, props);
+  };
+
+  return (
+    <div className={classNames.root}>
+      <div className={classNames.choiceFieldWrapper}>
+        <input
+          aria-label={ariaLabel}
+          id={id}
+          className={css(classNames.input, className)}
+          type="radio"
+          name={name}
+          disabled={disabled}
+          checked={checked}
+          required={required}
+          {...nativeProps}
+          onChange={onChange}
+          onFocus={onFocus}
+          onBlur={onBlur}
+        />
+        {onRenderField(props, defaultOnRenderField)}
+      </div>
+    </div>
+  );
+};
+ChoiceGroupOptionBase.displayName = 'ChoiceGroupOption';
