--- packages/react-next/src/components/ChoiceGroup/ChoiceGroupOption/index.ts
@@ -1,2 +1,0 @@
-export * from './ChoiceGroupOption';
-export * from './ChoiceGroupOption.types';
