--- packages/react-next/src/components/ChoiceGroup/index.ts
@@ -1,4 +1,0 @@
-export * from './ChoiceGroup';
-export * from './ChoiceGroup.base';
-export * from './ChoiceGroup.types';
-export * from './ChoiceGroupOption/index';
